Fix page breaks and move build script
